TorrentClient: extract TorrentSeeder class
