+++ src/main/java/net/yeputons/spbau/spring2016/torrent/client/TorrentClient.java
@@ -6,19 +6,18 @@ import net.yeputons.spbau.spring2016.torrent.protocol.*;
 
 import java.io.*;
 import java.net.InetSocketAddress;
-import java.net.ServerSocket;
-import java.net.Socket;
 import java.nio.ByteBuffer;
 import java.util.*;
-import java.util.stream.Collectors;
 
 public class TorrentClient implements Runnable {
     private final TorrentConnection tracker;
     private final ClientState state;
+    private TorrentSeeder seeder;
 
     public TorrentClient(TorrentConnection tracker, ClientState state) {
         this.tracker = tracker;
         this.state = state;
+        seeder = new TorrentSeeder(tracker, state);
     }
 
     @Override
@@ -38,42 +37,8 @@ public class TorrentClient implements Runnable {
             }).start();
         }
 
-        try (ServerSocket listener = new ServerSocket()) {
-            new Thread(() -> {
-                while (true) {
-                    try {
-                        listener.bind(new InetSocketAddress("0.0.0.0", 0));
-                        updateTracker(listener.getLocalPort());
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                        break;
-                    }
-                    try {
-                        // CHECKSTYLE.OFF: MagicNumber
-                        Thread.sleep(1000);
-                        // CHECKSTYLE.ON: MagicNumber
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }).start();
-
-            while (true) {
-                final Socket peer = listener.accept();
-                new Thread(() -> {
-                    try (DataInputStream in = new DataInputStream(peer.getInputStream());
-                         DataOutputStream out = new DataOutputStream(peer.getOutputStream())) {
-                        processPeer(peer, in, out);
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    } finally {
-                        try {
-                            peer.close();
-                        } catch (IOException ignored) {
-                        }
-                    }
-                }).start();
-            }
+        try {
+            seeder.start();
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -112,52 +77,6 @@ public class TorrentClient implements Runnable {
             } catch (InterruptedException e) {
                 break;
             }
-        }
-    }
-
-    private void updateTracker(int seedingPort) throws IOException {
-        tracker.makeRequest(new UpdateRequest(
-                seedingPort, state.getFiles().keySet().stream().collect(Collectors.toList())));
-    }
-
-    private void processPeer(final Socket peer, DataInputStream in, final DataOutputStream out)
-            throws IOException {
-        while (true) {
-            ClientRequest.readRequest(in).visit(new ClientRequestVisitor() {
-                @Override
-                public void accept(StatRequest r) throws IOException {
-                    synchronized (state) {
-                        FileDescription description = state.getFileDescription(r.getFileId());
-                        if (description != null) {
-                            r.answerTo(out,
-                                description.getDownloaded().stream().boxed().collect(Collectors.toList()));
-                        } else {
-                            r.answerTo(out, Collections.emptyList());
-                        }
-                    }
-                }
-
-                @Override
-                public void accept(GetRequest r) throws IOException {
-                    int fileId = r.getFileId();
-                    int partId = r.getPartId();
-                    FileDescription description;
-                    synchronized (state) {
-                        description = state.getFileDescription(fileId);
-                        if (description == null || !description.getDownloaded().get(r.getPartId())) {
-                            peer.close();
-                            return;
-                        }
-                    }
-                    RandomAccessFile file = state.getFile(r.getFileId());
-                    byte[] data = new byte[description.getPartSize(partId)];
-                    synchronized (file) {
-                        file.seek(description.getPartStart(partId));
-                        file.readFully(data);
-                    }
-                    r.answerTo(out, ByteBuffer.wrap(data));
-                }
-            });
         }
     }
 }
